Move Runner specific logic in a Runner specific target
