+++ Sources/Runner/Commands/Runner.swift
@@ -3,6 +3,8 @@ import Danger
 
 import Files
 import MarathonCore
+import Logger
+import RunnerLib
 
 func runDanger(logger: Logger) throws -> Void {
     // Pull in the JSON from Danger JS
